requirements.txt
beautifulsoup4==4.4.1
cssselect==0.9.1
lxml==3.4.4
requests==2.7.0
wheel==0.24.0
google==1.9.1
SQLAlchemy==1.0.12
MySQL-python==1.2.5
PySocks==1.5.6
unicodecsv==0.14.1
